Added more structure and options
