+++ k6-testResults.json
@@ -1,0 +1,1 @@
+{"root_group":{"name":"","path":"","id":"d41d8cd98f00b204e9800998ecf8427e","groups":[{"path":"::page_1 - https://staging.check-an-early-years-qualification.education.gov.uk/","id":"72a411fecafecffff3c08e1b9a37ace7","groups":[],"checks":[{"name":"has cookie 'auth-secret'","path":"::page_1 - https://staging.check-an-early-years-qualification.education.gov.uk/::has cookie 'auth-secret'","id":"ef9045648a7686d750f9195f3059003b","passes":30,"fails":0},{"id":"d776e2558b0e4966361e785e411c37a4","passes":30,"fails":0,"name":"'auth-secret' cookie has expected value","path":"::page_1 - https://staging.check-an-early-years-qualification.education.gov.uk/::'auth-secret' cookie has expected value"},{"path":"::page_1 - https://staging.check-an-early-years-qualification.education.gov.uk/::anti-forgery token has value","id":"aa1c1e36bc3eeca53174b65028ead59d","passes":30,"fails":0,"name":"anti-forgery token has value"},{"name":"request not challenged","path":"::page_1 - https://staging.check-an-early-years-qualification.education.gov.uk/::request not challenged","id":"c67727ac002ccddf29064407731a464e","passes":30,"fails":0},{"path":"::page_1 - https://staging.check-an-early-years-qualification.education.gov.uk/::status 200","id":"c202fb5f26284af6861135c81f4923db","passes":30,"fails":0,"name":"status 200"}],"name":"page_1 - https://staging.check-an-early-years-qualification.education.gov.uk/"},{"name":"page_2 - https://staging.check-an-early-years-qualification.education.gov.uk/questions/start-new","path":"::page_2 - https://staging.check-an-early-years-qualification.education.gov.uk/questions/start-new","id":"ec8837173a3310408434b6f2180b3888","groups":[],"checks":[{"id":"b2ffe2e65f47323329c195138dc1d28d","passes":30,"fails":0,"name":"request-verification token has value","path":"::page_2 - https://staging.check-an-early-years-qualification.education.gov.uk/questions/start-new::request-verification token has value"},{"name":"get status 200","path":"::page_2 - https://staging.check-an-early-years-qualification.education.gov.uk/questions/start-new::get status 200","id":"8f2746b739c41c20d4d97e21c5ef744b","passes":30,"fails":0},{"passes":30,"fails":0,"name":"get request not challenged","path":"::page_2 - https://staging.check-an-early-years-qualification.education.gov.uk/questions/start-new::get request not challenged","id":"609ca82924d040c5e92d9703383ad45c"},{"name":"post status 200","path":"::page_2 - https://staging.check-an-early-years-qualification.education.gov.uk/questions/start-new::post status 200","id":"487c436f85d2f4b731cdaf4042b09951","passes":120,"fails":0},{"id":"3d3f6f8d46d70c49eab1aa042c59a7d7","passes":120,"fails":0,"name":"post request not challenged","path":"::page_2 - https://staging.check-an-early-years-qualification.education.gov.uk/questions/start-new::post request not challenged"}]},{"name":"page_3 - https://staging.check-an-early-years-qualification.education.gov.uk/confirm-qualification/EYQ-224","path":"::page_3 - https://staging.check-an-early-years-qualification.education.gov.uk/confirm-qualification/EYQ-224","id":"81c2b6d1ceca99f468fcff154a5d3497","groups":[],"checks":[{"id":"ac438cf5d47ba8e7f0100c8a67f1d84f","passes":30,"fails":0,"name":"get status 200","path":"::page_3 - https://staging.check-an-early-years-qualification.education.gov.uk/confirm-qualification/EYQ-224::get status 200"},{"name":"get request not challenged","path":"::page_3 - https://staging.check-an-early-years-qualification.education.gov.uk/confirm-qualification/EYQ-224::get request not challenged","id":"0bb469b774295f7d74e1514434c6b824","passes":30,"fails":0}]},{"name":"page_4 - https://staging.check-an-early-years-qualification.education.gov.uk/confirm-qualification","path":"::page_4 - https://staging.check-an-early-years-qualification.education.gov.uk/confirm-qualification","id":"cca9e2a5791d60816addd5e7bea228af","groups":[],"checks":[{"name":"post status 200","path":"::page_4 - https://staging.check-an-early-years-qualification.education.gov.uk/confirm-qualification::post status 200","id":"1b4649ea50aa447d26470e2a08ed50b8","passes":60,"fails":0},{"name":"post request not challenged","path":"::page_4 - https://staging.check-an-early-years-qualification.education.gov.uk/confirm-qualification::post request not challenged","id":"f3b8d1b9b0f3fce134a8636f0299dfc7","passes":60,"fails":0}]},{"checks":[],"name":"page_5 - https://staging.check-an-early-years-qualification.education.gov.uk/qualifications/qualification-details/EYQ-224","path":"::page_5 - https://staging.check-an-early-years-qualification.education.gov.uk/qualifications/qualification-details/EYQ-224","id":"d4066498bbce0ed92d4091327e5a0395","groups":[]}],"checks":[]},"options":{"summaryTrendStats":["avg","min","med","max","p(90)","p(95)"],"summaryTimeUnit":"","noColor":false},"state":{"isStdOutTTY":true,"isStdErrTTY":true,"testRunDurationMs":42627.4133},"metrics":{"http_reqs":{"type":"counter","contains":"default","values":{"count":3900,"rate":91.4904212590352}},"vus":{"type":"gauge","contains":"default","values":{"value":4,"min":4,"max":10}},"http_req_duration{expected_response:true}":{"values":{"avg":96.72377082051294,"min":25.5385,"med":77.83685,"max":727.6668,"p(90)":169.22208,"p(95)":213.24854499999995},"type":"trend","contains":"time"},"group_duration":{"type":"trend","contains":"time","values":{"med":1463.8595500000001,"max":8163.7008,"p(90)":6678.92292,"p(95)":7207.223564999999,"avg":2531.7540533333336,"min":566.1514}},"http_req_connecting":{"contains":"time","values":{"avg":0.13313771794871793,"min":0,"med":0,"max":54.9379,"p(90)":0,"p(95)":0},"type":"trend"},"http_req_sending":{"type":"trend","contains":"time","values":{"avg":0.183565717948718,"min":0,"med":0,"max":4.0014,"p(90)":0.5628099999999999,"p(95)":0.5879}},"http_req_waiting":{"type":"trend","contains":"time","values":{"p(95)":106.33489499999999,"avg":67.90961417948716,"min":23.7557,"med":65.0469,"max":651.5403,"p(90)":96.54229}},"vus_max":{"contains":"default","values":{"value":10,"min":10,"max":10},"type":"gauge"},"http_req_blocked":{"type":"trend","contains":"time","values":{"avg":0.37215251282051276,"min":0,"med":0,"max":147.6753,"p(90)":0,"p(95)":0}},"http_req_failed":{"type":"rate","contains":"default","values":{"rate":0,"passes":0,"fails":3900}},"data_received":{"contains":"data","values":{"count":186496799,"rate":4375043.770249132},"type":"counter"},"http_req_tls_handshaking":{"type":"trend","contains":"time","values":{"avg":0.19269258974358977,"min":0,"med":0,"max":78.4896,"p(90)":0,"p(95)":0}},"data_sent":{"values":{"rate":23195.12077923809,"count":988748},"type":"counter","contains":"data"},"checks":{"type":"rate","contains":"default","values":{"rate":1,"passes":660,"fails":0}},"http_req_duration":{"type":"trend","contains":"time","values":{"avg":96.72377082051294,"min":25.5385,"med":77.83685,"max":727.6668,"p(90)":169.22208,"p(95)":213.24854499999995}},"http_req_receiving":{"contains":"time","values":{"avg":28.63059092307689,"min":0,"med":3.88885,"max":626.1094,"p(90)":95.8663,"p(95)":138.32033999999985},"type":"trend"},"iteration_duration":{"type":"trend","contains":"time","values":{"avg":13659.744336666663,"min":11611.1689,"med":13971.3903,"max":15062.6254,"p(90)":14759.5396,"p(95)":14911.739999999998}},"iterations":{"type":"counter","contains":"default","values":{"rate":0.7037724712233477,"count":30}}}}
