+++ tests/Dfe.EarlyYearsQualification.LoadTests/config/load.json
@@ -1,0 +1,44 @@
+{
+  "thresholds": {
+    "http_req_failed": [
+      "rate<0.01"
+    ],
+    "http_req_duration": [
+      "p(90) < 400",
+      "p(95) < 1000"
+    ]
+  },
+  "scenarios": {
+    "end_to_end": {
+      "executor": "ramping-vus",
+      "stages": [
+        {
+          "duration": "8m",
+          "target": 40
+        },
+        {
+          "duration": "8m",
+          "target": 80
+        },
+        {
+          "duration": "2m",
+          "target": 80
+        },
+        {
+          "duration": "2m",
+          "target": 40
+        },
+        {
+          "duration": "5m",
+          "target": 8
+        },
+        {
+          "duration": "5m",
+          "target": 0
+        }
+      ],
+      "gracefulStop": "5m",
+      "gracefulRampDown": "5m"
+    }
+  }
+}
