+++ tests/Dfe.EarlyYearsQualification.LoadTests/config/quick.json
@@ -1,0 +1,4 @@
+{
+  "vus": 5,
+  "duration": "15s"
+}
