+++ tests/Dfe.EarlyYearsQualification.LoadTests/tests/support/assurance.js
@@ -1,0 +1,9 @@
+
+export function getRequestVerificationTokenValue(response) {
+
+  const requestVerificationTokenInput = response.html().find('input[name=__RequestVerificationToken]');
+
+  const requestVerificationToken = requestVerificationTokenInput.attr('value');
+
+  return requestVerificationToken;
+}
